% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{colouRpop}
\alias{colouRpop}
\alias{colouRpop_palettes}
\title{colouRpop palettes}
\format{
An object of class \code{list} of length 11.
}
\usage{
colouRpop(palette = names(colouRpop_palettes), show_me = FALSE)

colouRpop_palettes
}
\arguments{
\item{palette}{A palette to choose.  See [colouRpop_palettes] for a list of applicable choices.}

\item{show_me}{Logical, if `TRUE` prints a `ggplot` figure of the palette}
}
\value{
A character vector of colours; if `show_me` is true, prints a plot
}
\description{
Plot palettes inspired by eye shadow palettes
}
\keyword{datasets}
